island(keystone #5, step 1): move cloud-client status store + secure store into island
Leaf moves ahead of relocating the CloudClient singleton itself:
- cloudClientStatus store -> island/src/stores (host @/stores/cloudClientStatus
  becomes a re-export shim); surfaced as toolkit.stores.cloudClientStatus.
- cloud secure store (MMKV KeyValueStore for the v2 refresh token) ->
  island/src/utils/cloudClient (host @/utils/cloudClient/MmkvSecureStore shim).
  react-native-mmkv is already an island dep, so this is adapter-free.
jest @mentra/island mock gains the store via requireActual (pure zustand).
